pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpc"
version = "0.1.0"
description = "Manifest Platform Core Suite — Python library suite for manifest-driven platforms"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "jsonschema>=4.18.0",
    "lark>=1.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
